@@ -11,20 +11,20 @@ exclude = [
 
 [project]
 name = "animetime"
-version = "0.1.6"
+version = "0.2"
 authors = [
- {name="swarthur", email=""}
+ {name="swarthur"}
 ]
 description = "A module to manage and track your favourites animes and series"
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.11"
 license = "MIT"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
-dependencies = ["animedata","tomli"]
+dependencies = ["animedata"]
 
 [project.urls]
 "Github" = "https://github.com/swarthur/animetime"
